+++ BackEnd/TgLab.Domain/DTOs/Bet/BetDTO.cs
@@ -6,9 +6,9 @@ namespace TgLab.Domain.DTOs.Bet
     {
         public int Id { get; set; }
         public string Currency { get; set; }
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public string Stage { get; set; }
-        public decimal Bounty { get; set; }
+        public double Bounty { get; set; }
         public DateTime Time { get; set; }
     }
 }
