+++ BackEnd/TgLab.Domain/DTOs/Wallet/DepositWalletedDTO.cs
@@ -3,6 +3,6 @@
     public class DepositWalletedDTO
     {
         public int Id { get; set; }
-        public decimal Amount {  get; set; }
+        public double Amount {  get; set; }
     }
 }
