+++ BackEnd/TgLab.Domain/DTOs/Wallet/WalletDTO.cs
@@ -5,7 +5,7 @@ namespace TgLab.Domain.DTOs.Wallet
     public class WalletDTO
     {
         public int Id { get; set; }
-        public decimal Balance { get; set; }
+        public double Balance { get; set; }
         public string Currency { get; set; }
     }
 }
