+++ BackEnd/TgLab.Domain/Interfaces/Wallet/IWalletService.cs
@@ -8,7 +8,7 @@ namespace TgLab.Domain.Interfaces.Wallet
         public Task Create(CreateWalletDTO dto, string userEmail);
         public Task Deposit(DepositWalletedDTO dto, string userEmail);
         public Task Withdraw(WithdrawWalletDTO dto, string userEmail);
-        public Task DecreaseBalance(WalletDb wallet, decimal amount);
-        public Task IncreaseBalance(WalletDb wallet, decimal bounty);
+        public Task DecreaseBalance(WalletDb wallet, double amount);
+        public Task IncreaseBalance(WalletDb wallet, double bounty);
     }
 }
